제_11회_산업통상자원부_공공데이터활용_빅데이터_분석_부문_공모전/분석코드/DLinear/requirements.txt
Python 3.8
Pytorch 1.9.0 
Cuda 11.1
Numpy 1.19.4
Matplotlib 3.3.3
Pandas 1.2.1